Implementing transfer learning for backwards solve
Forcibly represent the learning data as coming from same problem size as
transfer target or different sizes. Also permits a few varieties of
shuffles for the data as that can affect transfer learning
significantly.

Additionally, updated help to indicate file formats for outputs, add an
argument for single-process (debugging) and a script to run lots of
transfer learning variations and generate their plots.